@@ -18,9 +18,9 @@
     "db:studio": "prisma studio --browser none"
   },
   "dependencies": {
-    "@fontsource/anybody": "5.0.3",
-    "@fontsource/pt-mono": "5.0.3",
-    "@fontsource/pt-sans": "5.0.3",
+    "@fontsource/anybody": "5.0.5",
+    "@fontsource/pt-mono": "5.0.5",
+    "@fontsource/pt-sans": "5.0.5",
     "@prisma/client": "5.0.0",
     "@radix-ui/react-avatar": "1.0.3",
     "@radix-ui/react-hover-card": "1.0.6",
@@ -30,36 +30,36 @@
     "@remix-run/node": "^1.18.1",
     "@remix-run/react": "^1.18.1",
     "@remix-run/vercel": "^1.18.1",
-    "@vercel/node": "^2.10.3",
+    "@vercel/node": "^2.15.5",
     "class-variance-authority": "0.6.1",
-    "clsx": "1.2.1",
-    "isbot": "^3.6.8",
-    "lucide-react": "0.256.0",
+    "clsx": "2.0.0",
+    "isbot": "^3.6.12",
+    "lucide-react": "0.260.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "remix-utils": "6.6.0",
     "tailwind-merge": "1.13.2",
-    "tailwindcss": "3.3.2",
+    "tailwindcss": "3.3.3",
     "tailwindcss-animate": "1.0.6",
     "tiny-invariant": "1.3.1"
   },
   "devDependencies": {
-    "@ianvs/prettier-plugin-sort-imports": "4.0.2",
+    "@ianvs/prettier-plugin-sort-imports": "4.1.0",
     "@remix-run/dev": "^1.18.1",
     "@remix-run/eslint-config": "^1.18.1",
     "@remix-run/serve": "^1.18.1",
-    "@types/node": "20.3.3",
-    "@types/react": "^18.0.35",
-    "@types/react-dom": "^18.0.11",
-    "eslint": "^8.38.0",
+    "@types/node": "20.4.2",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "eslint": "^8.45.0",
     "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-prettier": "4.2.1",
-    "prettier": "2.8.8",
-    "prettier-plugin-tailwindcss": "0.3.0",
+    "eslint-plugin-prettier": "5.0.0",
+    "prettier": "3.0.0",
+    "prettier-plugin-tailwindcss": "0.4.1",
     "prisma": "5.0.0",
     "ts-node": "10.9.1",
     "tsconfig-paths": "4.2.0",
-    "typescript": "^5.0.4"
+    "typescript": "^5.1.6"
   },
   "engines": {
     "node": ">=14.0.0"
